tests/wpt/web-platform-tests/fullscreen/model/remove-last-manual.html

<!DOCTYPE html>
<title>Remove the last element on the fullscreen element stack</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="first">
    <div id="last"></div>
</div>
<script>
async_test(function(t)
{
    var first = document.getElementById("first");
    trusted_request(t, first);
    document.onfullscreenchange = t.step_func(function(event)
    {
        assert_equals(document.fullscreenElement, first);
        assert_equals(event.target, first);
        var last = document.getElementById("last");
        trusted_request(t, last);
        document.onfullscreenchange = t.step_func(function(event)
        {
            assert_equals(document.fullscreenElement, last);
            assert_equals(event.target, last);
            last.remove();
            document.onfullscreenchange = t.step_func_done(function(event)
            {
                assert_equals(document.fullscreenElement, first);
                assert_equals(event.target, document);
            });
        });
    });
});
</script>
